requirements.txt
kafka-python==2.0.2
tweepy==4.1.0
confluent-kafka==1.7.0
pymongo==3.12.0
python-dotenv==0.19.1
elasticsearch==7.15.1